Use multiple builds to conserve memory (#30)
A single Eleventy build runs out of memory while building a large Talk site, so this splits the build into a few smaller builds, which could be run in series or in parallel. The PR runs them in parallel for dev work but series for production, mostly because I haven't fully tested a parallel build with a large site like Penguin Watch or Planet Four.

The build scripts are (replace `build:` with `dev:` for local development)
- `build:api` builds `/api/collections`, `/api/project.json`, `/api/users` and `/api/tags`.
- `build:boards` builds `/boards`, `/api/boards` and `/api/discussions`.
- `build:site` builds `/collections`, `/recent`, `/users` and `/tags`.
- `build:subjects` builds `/api/subjects` and `/subjects`.

A couple of bugs are also fixed, so that subject discussions should now be linked up on subject pages and the project name should be passed properly to the layout templates.